converted, prep to ship
Improved and simplified Navigator window-dragging code.
Fixed an error with layering in drag and drop between Navigators.
Fixed an error with editing colors.
Added a toggle to the filter box to allow filter by test as well as text match.
Simplified the code to display the stack list.
Fixed an issue where mainstacks with templates for conversion to behaviors would show up twice in the option menu in the conversion dialog.
Added an option to the properties menu to open Size/Location Editor.
Updated the Size/Location Editor:
 -- The previous way of resizing (which involved following the sizing of a target object in real time) has been removed.
 -- Added a control to allow resizing/moving any control or set of controls: adjust the sides/corners, or move the objects
 -- Added a reset button to undo all changes made.
Added an option to the properties menu to open the colors editor.
Updated the Colors editor:
 -- Added backgrounds to the color sliders to make it clearer what edits what.
 -- Added a Choose button that opens the system color selector to put the value into the sliders.
 -- Reorganized the layout to make it clearer.
 -- Added tooltips to explain how color selection works.
 -- Added a color swatch to show what color is set on the sliders.
 -- Added a reset button to undo all changes made.
Changed preferences to save and restore them in an external text file. The format is fairly simple:
 -- All preferences are stored as a single line each.
 -- Each line is delimited into three items by numtochar(255).
 -- The first item is the property to be set.
 -- The second item is the name of the control that has the property -- this is a full name reference to the control, but not the long name. It uses the name of the mainstack instead of the filename. Thus the value can be used to set properties, but is portable from one installation to another.
 -- Values are urlencoded.
 -- In the one instance of a custompropertyset, the individual array values for the custompropertyset are urlencoded to be a single line, then the whole array is urlencoded to fit in the larger scheme.