6.3.1 Update Final
The Color Editor, the Size/Location Editor, the Property Editor, the Custom Property Editor, and the Command Panel all now display a description of the controls that will be affected, in the form of the name and ID of the primary control (the one from which color values etc. are displayed) followed by the number of additional controls selected for modification.

The Color Editor, the Size/Location Editor, the Property Editor, and the Command Panel, but *not* the Custom Property Editor, all now have a "Palette" button that opens them in separate palette windows. These windows can be minimized like standard Navigator windows, and the Property Editor and Command Panel can be resized as needed.

These new palettes will automatically update to reflect whatever the latest selection is in any open Navigator window. So you can select three controls, adjust their size in the Size/Location Editor, then select two different controls in a different Navigator window and the Size/Location Editor will immediately update to reflect that selection, and those two controls can be resized. 

It's also possible to change the primary control just by clicking a new control among the highlighted controls, which won't change the highlights, just the primary control.

The Preference dialog has a setting for "Editors as Palettes" that will cause the editors to always open as palettes. Whatever the setting, you can get the opposite by holding down the option/alt key while opening the editors.

This feature is new, and *very* powerful/dangerous. If you don't keep track of what controls Navigator wants to modify, you can do some serious damage really quickly. In my testing, the selection hasn't ever been wrong, but it's fairly easy not to realize that the selection has changed, even though all four palettes list what is being edited at the top of the palettes. BE CAREFUL WITH THIS FEATURE. But boy oh boy it's a lot of fun!

You can also have the palette open and then select the commands again and have the editor open in Navigator the way it has historically while the palette is still open. This is fine, albeit a little confusing.

The Stacks menu now has options for Show IDE Stacks and Show Script-Only Stacks. These affect what the menu displays, and also what the Stack List displays. 

The Cards menu now has options for Card List and Background List. These control what the menu displays and what shows in Navigator. So selecting Background List will cause Navigator to display a list of the backgrounds, and also the next time the menu is opened it will display a list of backgrounds.

Fixed a bug where the HTMLText caused the Property Editor to add a spurious line to its edit list.

Added a list of properties to display for Widgets (which previously displayed no properties). The list of properties is updatable in the Preferences. 

To-dos include: 
 -- a button in the palettes to switch back from the palette to the built-in editor
 -- making custom properties work in palette form -- should this be a separate (fifth) palette, or just an option in the properties palette?
 -- adding a Reset button to the properties/custom properties palette. There is no way to add a Reset button to the Command Panel.
 -- better documentation for the Command Panel.
 -- cleaning up the palettes -- each is just a stripped-down copy of Navigator, and some have not had as many unnecessary controls removed as others.
 -- better organization options for the palettes.
 -- option to lock a palette to a selection? Not sure.