6.6rc1
Navigator 6.6rc1 Release Notes

This is a major update to how commands work in Navigator, and continues to update how Navigator handles what container to target. Notes below.

=== NEW FEATURES ===

COMMAND BARS
Navigator now supports from 0-6 (configurable in preferences) Command Bars. These are similar in appearance to the Group Fold Bar, but on the right side of Navigator. Each bar can have a command associated with it, and clicking a selected control in the bar will apply that command to the currently selected controls. Click an unselected control in the Command Bar to select the control you click and apply the command to that one control.

This is intended as a more obvious way to apply different commands to a set of controls than the previous option/command/shift click options in the main list. Those options are still available and configurable in the preferences. That section of the preferences now includes the setting for how many command bars to display as well. There are no custom key combinations available for the command bars, except for control-clicking a command bar to select what command it should apply.

Each command bar has a tooltip that displays what command it will apply. To change the command, control-click or right-click the command bar. This displays a menu of the built-in commands, followed by any custom commands included in your Navigator Commands folder. The current command will have a checkmark; select any other command to change to that command. These settings are universal across all open copies of Navigator.

To define a new command, select New Command... at the top of the menu. That will open Navigator's Command Panel, where you can enter a new command. The command panel has a Save button. Once you save your command and give it a name, it will be set for that command panel -- and available as a custom command in general.

LIST CUSTOM COMMANDS NOW ALL USE THE COMMAND PANEL
In Preferences, list custom commands can now be defined using the Command Panel instead of just an Ask dialog, and have access to the pre-defined variables of the Command Panel -- tID for the current control, tCount for the number of controls selected, tIndex for the index of the current control in the list, etc. See the included Index Names.txt and Report Average Loc.txt for examples of how this can be used.

COMMAND PANEL CHANGES
The command panel has been updated. It now includes a tab for changing between commands and setting properties. It likely needs explaining that when setting properties, all that is needed is something like the following:

left=47
right = 89
visible= false

to set those three properties to the associated values for the selected controls. This can't be used as a custom command. The command panel now reflects this fact, hiding the "Save" button when showing the properties panel.

EDITING CUSTOM COMMANDS
There is no functionality for editing or removing custom commands. However, they are just text files in the Navigator Commands folder, so you can delete those files or edit them however you like.

STACK MENU BECOMES "WHICH TARGET" MENU -- INCLUDES RECENT TARGETS
The Stack menu is now the Target menu. The icon reflects the crosshair used for targeting a stack visually. Simply clicking the crosshair goes immediately to the mouseStack functionality. A slow click, or click-and-hold, displays the Stack menu as usual.

The recent targets have moved to the Stack menu. I upped the limit to 9, and internally I'm storing 99 -- let me know if there is any need/use case for this.

Recent targets are now universal across all copies of Navigator.

The Show IDE Stacks and Show Script-Only Stacks options should now stick from one launch of Navigator to the next -- but they're not preferences, so installing a new copy of Navigator will default them back to off.

OBSCURE, BUT NAVIGATOR HAS BETTER DEBUG FUNCTIONALITY
Navigator is a bear to debug. Over the years I can't count how many times I've typed:

put "got here"

to learn where my code is dying. That's a pain to find and disable when shipping Navigator, and more than once I've gotten a friendly note from someone saying that Navigator is triggering the message box. So I created a simple Debug command within Navigator. Here it is:

local sDebugLastCall
on debug
   -- exit debug
   if sDebugLastCall is empty or ticks() - sDebugLastCall > 120 then put empty
   put ticks() into sDebugLastCall
   put line -2 of the executionContexts into LC
   put ticks() & cr &  the long name of (item 1 of LC) & cr & item 2 to -1 of LC & cr & cr into R
   repeat with i = 1 to paramCount()
      put i & ":" && param(i) & cr & cr after R
   end repeat
   put R after message
end debug

Obviously, "debug" is easier to find and remove in scripts than an unterminated "put", but note that first line "-- exit debug"; as long as I uncomment that line, even if I miss a call to "debug", nothing will happen.

OBSCURE, BUT NAVIGATOR NOW HAS BUILT-IN IDE DEBUGGING CAPABILITY
Hold the option key while opening Navigator's Action menu and an option will be added at the bottom of the menu to debug Navigator/the IDE. When checked, breakpoints will work in IDE/Navigator stacks. Almost more importantly, unchecking it disables strict compilation.

A BUNCH OF SMALL BUG FIXES THAT I OUGHT TO DOCUMENT BETTER...
Yeah, sorry I don't have a list.

OUTSTANDING BUG
For several versions now Navigator's Drag and Drop has failed at dragging out of Navigator and into some stacks. I haven't figured out why I'm getting a mouseRelease message when I shouldn't, so I may have to re-write Navigator's drag and drop -- again :-/