7.0.1rc1
MultiTarget Windows

Navigator now supports multiple targets per window. You can still have as many Navigator windows as you like, but now each window can display the controls for any set of cards and groups.

Select multiple stacks in the stack list, then right-click and select browse controls, and Navigator will display all of the stacks in the same list.

In any display, select any combination of cards and groups. Right-click and select browse controls, and Navigator will display all of the containers in the same list.

In any command that sets the target for Navigator, hold down the Shift key to add to the target list instead of replacing it.

To remove a container from the list, right-click the container and select Stop Browsing from the popup menu. Navigator will remove the container from the display while retaining the others in the list.

Creating Navigator 7 required completely reworking how Navigator relayers controls, which resulted in several improvements. First, if you drag a control beneath a group, Navigator is now able to relayer the control properly, without forcing it into the group. You can still drag controls into and out of groups. And if you drag a grouped control to the bottom of the group, it will remain in the group.

The new relayering code is smart about whether to move or duplicate controls. Even if there are two representations of a container in the list, dragging a control from one to the other will simply relayer the control without duplicating it. Of course dragging a control from one container to another will duplicate it. 

Holding the option key reverses this behavior: if you hold the option key while dragging a control within its container, the control will be duplicated. And if you hold the option key while dragging a control to a different container, the control will be moved to the new location, and removed from the original.