7.1rc1
NAVIGATOR 7.1RC1 UPDATE NOTES

RELAYERING UPDATED (AGAIN)
This update completely overhauls relayering controls (again). I found multiple issues with the previous version of the code, starting with the fact that it was ugly and convoluted, and included a truly nasty hack to workaround for a very obscure bug I ran into in LiveCode that affected the display of controls that are relayered while editing a group. Since that's (nearly) inherent in the method I implemented for 7.0 to mostly avoid using relayerGroupedControls, I rewrote the whole thing for 7.1. In addition, the 7.0 code would relayer groups incorrectly (because it didn't take into account the number of controls in the group). 

In 7.1 I start editing and stop editing the target group for each individual control I relayer, and relayer each control twice, which is a bit distasteful, but it works a treat, and with much simpler code, so I'm happy.

A consequence of the way I'm relayering now is that relayering controls to the spot just below a group will always (one exception noted below) layer the control *out* of the group. In the initial release of Navigator 7 if the control was in the group, it would stay in the group, unless it was being dragged with another control that wasn't in the group. Now it's consistent. To get something to the bottom layer of a group, drag it into the group (anywhere but the bottom) and then drag the bottom control of the group above it. This leaves the same issue in the original Nav 7 update: it is currently impossible to drag controls into an empty group using Navigator. Bug noted here: https://github.com/gcanyon/navigator/issues/17

Exception: if a group is at the bottom of a card, dragging controls below that group will add the controls to that group. Bug noted here: https://github.com/gcanyon/navigator/issues/16

CUSTOM LIST STRINGS EDITOR
This update adds an editor for custom list strings. It shows a list of all the custom strings. You can select any string to edit it, and see a preview of what it will look like in Navigator. You can select multiple custom list strings to duplicate or delete.

BUGS FIXED
This update fixes the following bugs:
Setting card fold level broken: https://github.com/gcanyon/navigator/issues/13
Prefs file not set properly: https://github.com/gcanyon/navigator/issues/12

...and numerous other small issues. If you've emailed me about something recently, it's likely fixed by this update.